dedupe code between runway and cfngin vars + lookups
